Refactor code to move code from controller to service
